preprocess/requirements.txt
numpy<2
scipy
biopython
torch
fair-esm
h5py
